requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow